refactor: move writePromptToFile and resolveRoot to cli/util.go
