+++ internal/cli/root.go
@@ -2,8 +2,6 @@ package cli
 
 import (
 	"fmt"
-	"os"
-	"path/filepath"
 
 	"github.com/nqode/qode/internal/config"
 	"github.com/nqode/qode/internal/version"
@@ -93,39 +91,4 @@ func checkVersion(cmd *cobra.Command, _ []string) error {
 		return fmt.Errorf("project not initialised: run 'qode init'")
 	}
 	return version.CheckCompatibility(rootCmd.Version, cfg.QodeVersion)
-}
-
-// resolveRoot returns the effective project root, preferring the --root flag,
-// then the current working directory.
-func resolveRoot() (string, error) {
-	if flagRoot != "" {
-		return flagRoot, nil
-	}
-	wd, err := os.Getwd()
-	if err != nil {
-		return "", fmt.Errorf("cannot determine working directory: %w", err)
-	}
-	return wd, nil
-}
-
-// writePromptToFile atomically writes content to path, creating parent dirs as needed.
-// On template render error the caller should return before calling this.
-func writePromptToFile(path, content string) error {
-	if err := os.MkdirAll(filepath.Dir(path), 0700); err != nil {
-		return err
-	}
-	tmp, err := os.CreateTemp(filepath.Dir(path), ".qode-prompt-*")
-	if err != nil {
-		return err
-	}
-	tmpName := tmp.Name()
-	defer func() { _ = os.Remove(tmpName) }()
-	if _, err := tmp.WriteString(content); err != nil {
-		_ = tmp.Close()
-		return err
-	}
-	if err := tmp.Close(); err != nil {
-		return err
-	}
-	return os.Rename(tmpName, path)
 }
