+++ internal/cli/util.go
@@ -1,0 +1,42 @@
+package cli
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+)
+
+// resolveRoot returns the effective project root, preferring the --root flag,
+// then the current working directory.
+func resolveRoot() (string, error) {
+	if flagRoot != "" {
+		return flagRoot, nil
+	}
+	wd, err := os.Getwd()
+	if err != nil {
+		return "", fmt.Errorf("cannot determine working directory: %w", err)
+	}
+	return wd, nil
+}
+
+// writePromptToFile atomically writes content to path, creating parent dirs as needed.
+// On template render error the caller should return before calling this.
+func writePromptToFile(path, content string) error {
+	if err := os.MkdirAll(filepath.Dir(path), 0700); err != nil {
+		return err
+	}
+	tmp, err := os.CreateTemp(filepath.Dir(path), ".qode-prompt-*")
+	if err != nil {
+		return err
+	}
+	tmpName := tmp.Name()
+	defer func() { _ = os.Remove(tmpName) }()
+	if _, err := tmp.WriteString(content); err != nil {
+		_ = tmp.Close()
+		return err
+	}
+	if err := tmp.Close(); err != nil {
+		return err
+	}
+	return os.Rename(tmpName, path)
+}
